Some Minor Changes
