+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.9
 # Create app directory
-
+WORKDIR ./app
 # Install app dependencies
 COPY requirements.txt .
 
@@ -13,7 +13,7 @@ RUN  pip install -r requirements.txt
 
 
 # Bundle app source
-COPY . .
+COPY . ./app
 ENV FLASK_APP=app/run.py
 
 EXPOSE 5000
