+++ app/api.py
@@ -1,12 +1,22 @@
-
 from app.models import *
 from app import db
 
 import json
 import urllib
 import requests
+from marshmallow import Schema, fields
 
 
+# Schema For API
+class CarSchema(Schema):
+    id = fields.String()
+    year = fields.Integer()
+    make = fields.String()
+    created_at = fields.String()
+    updated_at = fields.String()
+
+
+# Function for Celery task to fetch data
 
 def fetch_from_api():
     where = urllib.parse.quote_plus("""
