+++ app/make_celery.py
@@ -5,6 +5,7 @@ def make_celery(app):
     celery = Celery(app.import_name)
     celery.conf.update(app.config["CELERY_CONFIG"])
     celery.conf.update(beat_schedule=app.config["CELERYBEAT_SCHEDULE"])
+
     class ContextTask(celery.Task):
         def __call__(self, *args, **kwargs):
             with app.app_context():
