+++ app/run.py
@@ -2,11 +2,4 @@ from app import app
 
 if __name__ == "__main__":
     app.run(host='0.0.0.0')
-# creating schema of api by extending it with marshmallow_Schema
-
-
-# Api Configurations
-
-
-# fetching data from api and storing in Database
-# created api to display fetched data (with serailizer using marshmallow schema)
+#
