+++ app/templates/index.html
@@ -7,8 +7,8 @@
     <title>Document</title>
 </head>
 <body>
-    This is Home
-    <h1>{{message}}</h1>
-    {{data}}
+This is Home
+<h1>{{ message }}</h1>
+{{ data }}
 </body>
 </html>
