+++ docker-compose.yml
@@ -1,47 +1,54 @@
 services:
-#    redis:
-#        image: 'bitnami/redis:latest'
-#        environment:
-#                - ALLOW_EMPTY_PASSWORD=yes
-#        container_name: redis3
-#        ports:
-#            - '6379:6379'
+  #    redis:
+  #        image: 'bitnami/redis:latest'
+  #        environment:
+  #                - ALLOW_EMPTY_PASSWORD=yes
+  #        container_name: redis3
+  #        ports:
+  #            - '6379:6379'
 
-    web:
-        build: .
-        container_name: web3
-        ports:
-            - "5000:5000"
+  web:
+    container_name: web3
+    build: .
+    ports:
+      - "5000:5000"
+    volumes:
+      - .:/app
+    depends_on:
+      - redis
 
-        depends_on:
-            - redis
+  redis:
+    image: redis
+    container_name: redis2
+    volumes:
+      - .:/app
 
-    redis:
-        image: redis
-        container_name: redis2
+  celery:
+    image: celery:3.1
+    container_name: celery7
+    restart: unless-stopped
+    working_dir: /app
+    volumes:
+      - .:/app
+    command: celery --app app.task.celery worker --loglevel=info
 
-    celery:
-        image: celery:3.1
-        container_name: celery7
-        restart: unless-stopped
-        build: .
-        command: celery --app app.task.celery worker --loglevel=info
+    depends_on:
+      - redis
+      - web
+    links:
+      - redis
+  CeleryScheduler:
+    image: celery:3.1
+    container_name: celery-schedule
+    working_dir: /app
+    restart: unless-stopped
+    volumes:
+      - .:/app
+    command: celery --app app.task.celery beat
 
-        depends_on:
-            - redis
-            - web
-        links:
-            - redis
-    CeleryScheduler:
-        image: celery:3.1
-        container_name: celery-schedule
-        restart: unless-stopped
-        build: .
-        command: celery --app app.task.celery beat
-
-        depends_on:
-            - redis
-            - web
-            - celery
-        links:
-            - redis
+    depends_on:
+      - redis
+      - web
+      - celery
+    links:
+      - redis
